Added Movie Cover images and Gallery
Added a Gallery Feature. Gallery is an example of bootstrap's carousel. It dynamically displays cover images and titles of all movies in the database. Found resizing images a little difficult so I left them full size for now.

Added a new field to the database: CoverURL.

Added cover images to the popovers in the RpMovies page